requirements.txt
altgraph==0.17.2
asgiref==3.5.2
auto-py-to-exe==2.20.1
backports.zoneinfo==0.2.1
bottle==0.12.21
bottle-websocket==0.2.9
cffi==1.15.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.5
django-cors-headers==3.13.0
django-filter==21.1
django-import-export==2.8.0
# django-simple-history @ file:///C:/Users/%D0%9E%D0%BB%D1%8C%D0%B3%D0%B0/Documents/tasktracker_dj/django-simple-history-3.1.1.tar.gz
djangorestframework==3.13.1
Eel==0.14.0
et-xmlfile==1.1.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
pefile==2022.5.30
pycparser==2.21
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.6
pyparsing==3.0.9
pytz==2022.1
pywin32-ctypes==0.2.0
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
whichcraft==0.6.1
xlrd==2.0.1
xlwt==1.3.0
zope.event==4.5.0
zope.interface==5.4.0
